refactor(controller.LottoController) : generateLottoList() 메서드를 LottoService로 이동
[피드백] 비즈니스 로직과 UI 로직을 분리한다.
를 반영하여 generateLottoList() 메서드는
비즈니스 로직이라 판단했기 때문에
LottoController에서 LottoSerivice 로
메서드를 이동시켰습니다.
